requirements.txt
torch
transformers
sentencepiece
datasets
rouge
wandb
accelerate
psycopg2
faiss-cpu
onnx
onnxruntime
jax
flax